chore(permission-sets)!: replace account ID with account name in `aws_ssoadmin_account_assignment` (#219)
Previously, resource keys used the account ID, which caused issues
during the plan phase when account IDs were not yet known (when accounts
were being created in the same run).

This change updates the resource key to use the account name instead, as this
does not have to be a computed value.

Updates include:
- Adjusted the `var.assignments` to be a list of objects
- Updated examples in the README to reflect the new structure
- Enhanced the variable description for clarity

BREAKING CHANGE: this commit changes the structure of
`var.aws_sso_permission_sets` in the root module and `var.assignments`
in the `permission-set` sub-module. In addition the
`aws_ssoadmin_account_assignment` resources will be recreated using the
new format.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>